Reduce number of messages sent and fix timeout
Added a feature that makes it possible to get the message history
for a channel, making it easier to find missing functionality and
responses. Right now, it's all just saved to a text file in the
current directory.

Added the possibility of debugging the bot interactively from the
command line where the bot is running from.