Improving the markov chain functionality
Added a new function in markov-generator that lets me check if a
given word is in the startWords array in markov-generator. Made it
possible to give makeChain a word to start from. I'll publish a
fork or something later.

Added a markov chain per user with over 150 messages in the channel,
in addition to a fuckton of chains based on all the messages on the
channel overall.

Created a few helper scripts to make it easier to 1) add data to
the database and 2) get data from the database.