Made many changes, some are important others are not
Added my own markov chain generator, because third party
libraries were too slow for the amount of messages I want
to be able to support.

Because I'm still running this from my own computer, I want
the script to not fail if it lose internet connection for an
hour. All listeners I've tried in discord.js have not worked
properly for this purpose, eventually throwing errors and
ending the process. I tried relying on forever for a few weeks,
but that too was too unreliable. So, here I tried making a new
function, runForever(), to make us just ignore exceptions and
keep trying to restart.

The codebase is a mess right now. Maybe I'll clean it up later.