Security fix and removing of Markov Chain logic
Google-translate-api was relying on safe-eval, which had a sandbox
escape vulnerability. As this package is no longer supported, I changed
to a supported fork of the package instead.

Moved the logic for creating and generating Markov Chains out of this
project, and into a new repository and package (easy-markov-chain).

Removed temporary code from fillDatabase.js.