Changed creation of simple messages and responses
Simple messages and simple message responses are now moved into separate
files that are far easier to edit for users with limited knowledge and
understanding of the logic of the bot. Added three new files for this:
simple-messages.yml, simple-message-replies.yml and
simple-fuzzy-messages.yml.

Moved several message-response-mappings from the msg, simpleMsg and
simpleMsgReply arrays into the YAML-files.

Moved the Markov Chain generation from the message event listener, and
into a responseStrategy in the responseStrategies array.

In general, cleaned the code a little here and there.