Added online check
If the bot is being ran from a computer with an internet connection that
fails sometimes, we will now check if we are online by using the
is-online package. If the internet connection is lost for a few minutes,
we exit the process. Maybe there will be another strategy for handling
this later.